refactor(tasks): move locomotion DR provider [skip ci]
