+++ src/unilab/tasks/locomotion/go2/joystick.py
@@ -13,7 +13,6 @@ from unilab.base.scene import SceneCfg
 from unilab.dtype_config import get_global_dtype
 from unilab.envs.locomotion.common import rewards
 from unilab.envs.locomotion.common.base import Sensor
-from unilab.envs.locomotion.common.dr_provider import LocomotionDRProvider
 from unilab.envs.locomotion.common.rewards import RewardContext
 from unilab.envs.locomotion.common.terrain_spawn import (
     TerrainCurriculumCfg,
@@ -25,6 +24,7 @@ from unilab.envs.manager_based_rl_env import (
 )
 from unilab.tasks.locomotion.common.commands import Commands
 from unilab.tasks.locomotion.common.domain_rand import DomainRandConfig
+from unilab.tasks.locomotion.common.dr_provider import LocomotionDRProvider
 from unilab.tasks.locomotion.go2.base import Go2BaseCfg, Go2BaseEnv
